pyproject.toml
[tool.poetry]
name = "simple-task-manager"
version = "0.1.0"
description = ""
authors = ["deymonster <deymonster@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
drf-yasg = "^1.21.8"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
